gwpy.testing: renamed old tests module
since this provides testing utilities, and not any actual tests
