+++ gwpy/frequencyseries/tests/test_hist.py
@@ -30,7 +30,7 @@ from matplotlib import rc_context
 from astropy import units
 
 from ...segments import Segment
-from ...tests import utils
+from ...testing import utils
 from ...types.tests.test_array2d import TestArray2D as _TestArray2D
 from .. import SpectralVariance
