+++ gwpy/io/tests/test_cache.py
@@ -30,7 +30,7 @@ import numpy
 import pytest
 
 from ...segments import (Segment, SegmentList)
-from ...tests.utils import (skip_missing_dependency, TemporaryFilename)
+from ...testing.utils import (skip_missing_dependency, TemporaryFilename)
 from .. import cache as io_cache
 
 __author__ = 'Duncan Macleod <duncan.macleod@ligo.org>'
