+++ gwpy/testing/__init__.py
@@ -16,8 +16,5 @@
 # You should have received a copy of the GNU General Public License
 # along with GWpy.  If not, see <http://www.gnu.org/licenses/>.
 
-"""Unit tests for GWpy
+"""Testing utilities for GWpy
 """
-
-from matplotlib import use
-use('agg')
