+++ gwpy/testing/data/H1-LDAS_STRAIN-968654552-10.xml.gz
