+++ gwpy/testing/data/HLV-HW100916-968654552-1.gwf
