+++ gwpy/testing/data/X1-GWPY_TEST_SEGMENTS-0-10.xml.gz
