+++ gwpy/testing/mocks.py
