+++ gwpy/testing/utils.py
